Remove .git from gitlab dependency to fix redirect loop.
When renovate attempts to process the dependencies for lyrebird, it attempts to
go to: https://gitlab.com/yawning/edwards25519-extra.git?go-get=1. For some
reason, that produces a redirect loop in gitlab, and it causes the renovate run
to fail after a redirect limit is reached. You can see the redirect yourself by
visiting that URL in your browser.

With this change, by simply removing the .git at the end, renovate would visit
https://gitlab.com/yawning/edwards25519-extra?go-get=1 and that does not produce
a redirect loop.